requirements.txt
asgiref==3.5.2
asyncio==3.4.3
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
contextlib2==21.6.0
cryptography==37.0.4
Django==4.1
django-async-orm==0.1.10
django-cors-headers==3.13.0
django-ninja==0.19.1
django-ninja-extra==0.15.4
django-ninja-jwt==5.2.0
dnspython==2.2.1
email-validator==1.2.1
idna==3.3
imagekitio==2.2.8
injector==0.19.0
ninja-schema==0.12.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
